UserCRUD/src/test/java/com/nttdata/usercrud/UserControllerTest.java
package com.nttdata.usercrud;

import com.nttdata.usercrud.controller.UserController;
import com.nttdata.usercrud.service.UserService;
import com.nttdata.usercrud.util.DataTest;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(UserController.class)
public class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @Test
    void testGetUserById() throws Exception {
        when(userService.getUserById(1)).thenReturn(DataTest.saverUser01().orElseThrow());
        mockMvc.perform(get("/user/1").contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.id").value(1))
                .andExpect(jsonPath("$.name").value("User 01"))
                .andExpect(jsonPath("$.email").value("user01@test.com"))
                .andExpect(jsonPath("$.age").value(20));

        verify(userService).getUserById(1);
    }



}
